% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cities}
\alias{cities}
\title{Point layer of the three largest cities in Israel}
\format{
A \code{sf} POINT layer with 3 features and 1 attribute:
\describe{
  \item{name}{Town name}
}
}
\usage{
cities
}
\description{
A \code{sf} POINT layer of the three largest cities in Israel: Jerusalem, Tel-Aviv and Haifa.
}
\examples{
plot(cities)
}
\keyword{datasets}
