% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dem}
\alias{dem}
\title{Small Digital Elevation Model}
\format{A \code{stars} object with 1 attribute:
\describe{
  \item{elevation}{Elevation above sea level, in meters}
}}
\usage{
dem
}
\description{
A \code{stars} object representing a 13*11 Digital Elevation Model (DEM)
}
\keyword{datasets}
