% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{line}
\alias{line}
\title{Sample network dataset: lines}
\format{An \code{sf} object}
\usage{
line
}
\description{
An \code{sf} object based on the \code{edge_table} sample dataset from pgRouting 2.6 tutorial
}
\references{
\url{https://docs.pgrouting.org/2.6/en/sampledata.html}
}
\keyword{datasets}
