% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nngeo-package.R
\docType{package}
\name{nngeo-package}
\alias{nngeo}
\alias{nngeo-package}
\title{nngeo: k-Nearest Neighbor Join for Spatial Data}
\description{
K-nearest neighbor search for projected and non-projected 'sf' spatial layers. Nearest neighbor search uses (1) C code from 'GeographicLib' for lon-lat point layers, (2) function knn() from package 'nabor' for projected point layers, or (3) function st_distance() from package 'sf' for line or polygon layers. The package also includes several other utility functions for spatial analysis.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://michaeldorman.github.io/nngeo/}
  \item \url{https://github.com/michaeldorman/nngeo/}
  \item Report bugs at \url{https://github.com/michaeldorman/nngeo/issues/}
}

}
\author{
\strong{Maintainer}: Michael Dorman \email{dorman@post.bgu.ac.il}

Other contributors:
\itemize{
  \item Johnathan Rush \email{johnathan.rush@mssm.edu} [contributor]
  \item Ian Hough \email{hough@post.bgu.ac.il} [contributor]
  \item Dominic Russel \email{domrussel@gmail.com} [contributor]
  \item Luigi Ranghetti \email{luigi.ranghetti@gmail.com} [contributor]
  \item Charles F.F Karney (Author of included C code from 'GeographicLib' for geodesic distance) [contributor, copyright holder]
}

}
\keyword{internal}
