% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pnt}
\alias{pnt}
\title{Sample network dataset: points}
\format{
An \code{sf} object
}
\usage{
pnt
}
\description{
An \code{sf} object based on the \code{pointsOfInterest} sample dataset from pgRouting 2.6 tutorial
}
\examples{
plot(pnt)
}
\references{
\url{https://docs.pgrouting.org/2.6/en/sampledata.html}
}
\keyword{datasets}
