% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{towns}
\alias{towns}
\title{Point layer of towns in Israel}
\format{A \code{sf} POINT layer with 93 features and 1 attribute:
\describe{
  \item{name}{Town name}
}}
\usage{
towns
}
\description{
A \code{sf} POINT layer of all towns in Israel whose name starts with the letter "A".
}
\keyword{datasets}
