% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{water}
\alias{water}
\title{Polygonal layer of water bodies in Israel}
\format{A \code{sf} POLYGON layer with 4 features and 1 attribute:
\describe{
  \item{name}{Water body name}
}}
\usage{
water
}
\description{
A \code{sf} POLYGON layer of the four large water bodies in Israel:
\itemize{
\item{Mediterranean Sea}
\item{Red Sea}
\item{Sea of Galilee}
\item{Dead Sea}
}
}
\keyword{datasets}
